Split Feedback page into components
