+++ ui/src/components/layout/alerts/ErrorAlert.tsx
@@ -6,15 +6,19 @@ import {
 import {XCircle} from "lucide-react";
 import React from "react";
 
-export function ErrorAlert() {
+export function ErrorAlert({ isActive } : { isActive:any }) {
   return (
-    <Alert variant="destructive" className="bg-red-100">
-      <XCircle className="h-4 w-4" />
-      <AlertTitle>Error</AlertTitle>
-      <AlertDescription className="text-base text-red-500">
-        Email feedback was unsuccessful. Please try again.
-      </AlertDescription>
-    </Alert>
+    <section className="errorAlert">
+      {isActive ? (
+        <Alert variant="destructive" className="bg-red-100">
+          <XCircle className="h-4 w-4" />
+          <AlertTitle>Error</AlertTitle>
+          <AlertDescription className="text-base text-red-500">
+            Email feedback was unsuccessful. Please try again.
+          </AlertDescription>
+        </Alert>
+      ) : null }
+    </section>
   )}
 
 export default ErrorAlert;
