+++ ui/tests/app/feedback/page.test.tsx
@@ -1,6 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import Feedback from "@/app/feedback/page";
 import userEvent from '@testing-library/user-event'
+import {wait} from "@testing-library/user-event/utils/misc/wait";
 
 describe('Feedback', () => {
   it('should render the Feedback form', async () => {
@@ -17,14 +18,9 @@ describe('Feedback', () => {
   it('should see the success alert after submission', async() => {
     const user = userEvent.setup()
     render(<Feedback />);
-    expect(screen.getByText('Feedback Type:')).toBeInTheDocument();
-    expect(screen.getByText('Your Name (Optional):')).toBeInTheDocument();
-    expect(screen.getByText('Email Address:')).toBeInTheDocument();
-    expect(screen.getByText('Your Feedback:')).toBeInTheDocument();
     await user.type(screen.getByText('Feedback Type:'), 'General');
     await user.type(screen.getByText('Your Name (Optional):'), 'John Doe');
-    await user.type(screen.getByText('Email Address:'), 'test@example.com');
-    await user.type(screen.getByText('Your Feedback:'), 'This is a test feedback message.');
+    await user.type(screen.getByText('Your Feedback:'), 'This is a to test if the success alert shows.');
     await user.click(screen.getByRole('button', {name: 'Submit'}));
 
     // Wait for the success alert to appear
@@ -35,14 +31,9 @@ describe('Feedback', () => {
   it('should see the error alert after submission', async() => {
     const user = userEvent.setup()
     render(<Feedback />);
-    expect(screen.getByText('Feedback Type:')).toBeInTheDocument();
-    expect(screen.getByText('Your Name (Optional):')).toBeInTheDocument();
-    expect(screen.getByText('Email Address:')).toBeInTheDocument();
-    expect(screen.getByText('Your Feedback:')).toBeInTheDocument();
     await user.type(screen.getByText('Feedback Type:'), 'General');
     await user.type(screen.getByText('Your Name (Optional):'), 'John Doe');
-    await user.type(screen.getByText('Email Address:'), 'test@example.com');
-    await user.type(screen.getByText('Your Feedback:'), 'Short message.');
+    await user.type(screen.getByText('Your Feedback:'), 'Short message to test if the error alert shows.');
     await user.click(screen.getByRole('button', {name: 'Submit'}));
 
     // Wait for the error alert to appear
